RF: Split out conversions
