+++ fendermint/app/src/app.rs
@@ -1,11 +1,9 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
 use std::future::Future;
-use std::num::NonZeroU32;
 use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
 
-use anyhow::anyhow;
 use async_trait::async_trait;
 use cid::Cid;
 use fendermint_abci::Application;
@@ -18,33 +16,22 @@ use fendermint_vm_interpreter::bytes::{
 };
 use fendermint_vm_interpreter::chain::{ChainMessageApplyRet, IllegalMessage};
 use fendermint_vm_interpreter::fvm::{
-    FvmApplyRet, FvmCheckRet, FvmCheckState, FvmExecState, FvmGenesisState, FvmQueryRet,
-    FvmQueryState,
+    FvmApplyRet, FvmCheckState, FvmExecState, FvmGenesisState, FvmQueryState,
 };
 use fendermint_vm_interpreter::signed::InvalidSignature;
-use fendermint_vm_interpreter::{CheckInterpreter, Interpreter, QueryInterpreter, Timestamp};
+use fendermint_vm_interpreter::{CheckInterpreter, Interpreter, QueryInterpreter};
 use fvm_ipld_blockstore::Blockstore;
 use fvm_shared::econ::TokenAmount;
-use fvm_shared::error::ExitCode;
-use fvm_shared::event::StampedEvent;
 use fvm_shared::version::NetworkVersion;
 use serde::{Deserialize, Serialize};
 use tendermint::abci::request::CheckTxKind;
-use tendermint::abci::{request, response, Code, Event, EventAttribute};
+use tendermint::abci::{request, response};
 use tendermint::block::Height;
 
+use crate::tmconv::*;
+use crate::BlockHeight;
+
 const VERSION: &str = env!("CARGO_PKG_VERSION");
-
-/// IPLD encoding of data types we know we must be able to encode.
-macro_rules! must_encode {
-    ($var:ident) => {
-        fvm_ipld_encoding::to_vec(&$var)
-            .unwrap_or_else(|e| panic!("error encoding {}: {}", stringify!($var), e))
-    };
-}
-
-// Different type from `ChainEpoch` just because we might use epoch in a more traditional sense for checkpointing.
-pub type BlockHeight = u64;
 
 #[derive(Serialize)]
 #[repr(u8)]
@@ -54,7 +41,7 @@ pub enum AppStoreKey {
 
 // TODO: What range should we use for our own error codes? Should we shift FVM errors?
 #[repr(u32)]
-enum AppError {
+pub enum AppError {
     /// Failed to deserialize the transaction.
     InvalidEncoding = 51,
     /// Failed to validate the user signature.
@@ -377,14 +364,7 @@ where
         let db = self.clone_db();
         let state = self.committed_state();
         let height = request.header.height.into();
-        let timestamp = Timestamp(
-            request
-                .header
-                .time
-                .unix_timestamp()
-                .try_into()
-                .expect("negative timestamp"),
-        );
+        let timestamp = to_timestamp(request.header.time);
 
         let state = FvmExecState::new(
             db,
@@ -458,190 +438,4 @@ where
             retain_height: Default::default(),
         }
     }
-}
-
-/// Response to delivery where the input was blatantly invalid.
-/// This indicates that the validator who made the block was Byzantine.
-fn invalid_deliver_tx(err: AppError, description: String) -> response::DeliverTx {
-    response::DeliverTx {
-        code: Code::Err(NonZeroU32::try_from(err as u32).expect("error codes are non-zero")),
-        info: description,
-        ..Default::default()
-    }
-}
-
-/// Response to checks where the input was blatantly invalid.
-/// This indicates that the user who sent the transaction is either attacking or has a faulty client.
-fn invalid_check_tx(err: AppError, description: String) -> response::CheckTx {
-    response::CheckTx {
-        code: Code::Err(NonZeroU32::try_from(err as u32).expect("error codes are non-zero")),
-        info: description,
-        ..Default::default()
-    }
-}
-
-/// Response to queries where the input was blatantly invalid.
-fn invalid_query(err: AppError, description: String) -> response::Query {
-    response::Query {
-        code: Code::Err(NonZeroU32::try_from(err as u32).expect("error codes are non-zero")),
-        info: description,
-        ..Default::default()
-    }
-}
-
-fn to_deliver_tx(ret: FvmApplyRet) -> response::DeliverTx {
-    let receipt = ret.apply_ret.msg_receipt;
-    let code = to_code(receipt.exit_code);
-
-    // Based on the sanity check in the `DefaultExecutor`.
-    // gas_cost = gas_fee_cap * gas_limit; this is how much the account is charged up front.
-    // &base_fee_burn + &over_estimation_burn + &refund + &miner_tip == gas_cost
-    // But that's in tokens. I guess the closes to what we want is the limit.
-    let gas_wanted: i64 = ret.gas_limit.try_into().unwrap_or(i64::MAX);
-    let gas_used: i64 = receipt.gas_used.try_into().unwrap_or(i64::MAX);
-
-    let data = receipt.return_data.to_vec().into();
-    let events = to_events("message", ret.apply_ret.events);
-
-    response::DeliverTx {
-        code,
-        data,
-        log: Default::default(),
-        info: Default::default(),
-        gas_wanted,
-        gas_used,
-        events,
-        codespace: Default::default(),
-    }
-}
-
-fn to_check_tx(ret: FvmCheckRet) -> response::CheckTx {
-    response::CheckTx {
-        code: to_code(ret.exit_code),
-        gas_wanted: ret.gas_limit.try_into().unwrap_or(i64::MAX),
-        sender: ret.sender.to_string(),
-        ..Default::default()
-    }
-}
-
-fn to_code(exit_code: ExitCode) -> Code {
-    if exit_code.is_success() {
-        Code::Ok
-    } else {
-        Code::Err(NonZeroU32::try_from(exit_code.value()).expect("error codes are non-zero"))
-    }
-}
-
-/// Map the return values from epoch boundary operations to validator updates.
-///
-/// (Currently just a placeholder).
-fn to_end_block(_ret: ()) -> response::EndBlock {
-    response::EndBlock {
-        validator_updates: Vec::new(),
-        consensus_param_updates: None,
-        events: Vec::new(),
-    }
-}
-
-/// Map the return values from cron operations.
-///
-/// (Currently just a placeholder).
-fn to_begin_block(ret: FvmApplyRet) -> response::BeginBlock {
-    let events = to_events("begin", ret.apply_ret.events);
-
-    response::BeginBlock { events }
-}
-
-/// Convert events to key-value pairs.
-fn to_events(kind: &str, stamped_events: Vec<StampedEvent>) -> Vec<Event> {
-    stamped_events
-        .into_iter()
-        .map(|se| {
-            let mut attrs = Vec::new();
-
-            attrs.push(EventAttribute {
-                key: "emitter".to_string(),
-                value: se.emitter.to_string(),
-                index: true,
-            });
-
-            for e in se.event.entries {
-                attrs.push(EventAttribute {
-                    key: e.key,
-                    value: hex::encode(e.value),
-                    index: !e.flags.is_empty(),
-                });
-            }
-
-            Event::new(kind.to_string(), attrs)
-        })
-        .collect()
-}
-
-/// Map to query results.
-fn to_query(ret: FvmQueryRet, block_height: BlockHeight) -> response::Query {
-    let exit_code = match ret {
-        FvmQueryRet::Ipld(None) | FvmQueryRet::ActorState(None) => ExitCode::USR_NOT_FOUND,
-        FvmQueryRet::Ipld(_) | FvmQueryRet::ActorState(_) => ExitCode::OK,
-    };
-
-    // The return value has a `key` field which is supposed to be set to the data matched.
-    // Although at this point I don't have access to the input like the CID looked up,
-    // but I assume the query sender has. Rather than repeat everything, I'll add the key
-    // where it gives some extra information, like the actor ID, just to keep this option visible.
-    let (key, value) = match ret {
-        FvmQueryRet::Ipld(None) | FvmQueryRet::ActorState(None) => (Vec::new(), Vec::new()),
-        FvmQueryRet::Ipld(Some(bz)) => (Vec::new(), bz),
-        FvmQueryRet::ActorState(Some(x)) => {
-            let (id, st) = *x;
-            let k = must_encode!(id);
-            let v = must_encode!(st);
-            (k, v)
-        }
-    };
-
-    // The height here is the height of the block that was committed, not in which the app hash appeared,
-    // so according to Tendermint docstrings we need to return plus one.
-    let height = tendermint::block::Height::try_from(block_height + 1).expect("height too big");
-
-    response::Query {
-        code: to_code(exit_code),
-        key: key.into(),
-        value: value.into(),
-        height,
-        ..Default::default()
-    }
-}
-
-/// Parse the initial genesis either as JSON or CBOR.
-fn parse_genesis(bytes: &[u8]) -> anyhow::Result<Genesis> {
-    try_parse_genesis_json(bytes).or_else(|e1| {
-        try_parse_genesis_cbor(bytes)
-            .map_err(|e2| anyhow!("failed to deserialize genesis as JSON or CBOR: {e1}; {e2}"))
-    })
-}
-
-fn try_parse_genesis_json(bytes: &[u8]) -> anyhow::Result<Genesis> {
-    let json = String::from_utf8(bytes.to_vec())?;
-    let genesis = serde_json::from_str(&json)?;
-    Ok(genesis)
-}
-
-fn try_parse_genesis_cbor(bytes: &[u8]) -> anyhow::Result<Genesis> {
-    let genesis = fvm_ipld_encoding::from_slice(bytes)?;
-    Ok(genesis)
-}
-
-/// Project Genesis validators to Tendermint.
-fn genesis_validators(genesis: &Genesis) -> anyhow::Result<Vec<tendermint::validator::Update>> {
-    let mut updates = vec![];
-    for v in genesis.validators.iter() {
-        let bz = v.public_key.0.serialize();
-        let key = k256::ecdsa::VerifyingKey::from_sec1_bytes(&bz)?;
-        updates.push(tendermint::validator::Update {
-            pub_key: tendermint::public_key::PublicKey::Secp256k1(key),
-            power: tendermint::vote::Power::try_from(v.power.0)?,
-        });
-    }
-    Ok(updates)
 }
